feat: decouples serializer from Maze
The serializer.py module contained a dump() and load() functions. These
functions dependend on the Maze class from the maze.py module. This
will create a cyclical import error when importing the serializer into
the maze.py module.

Modifies the serializer.py methods to work with the building blocks from
Maze. The functions are renamed to dump_squares() and load_squares() to
better convey their purpose.